requirements.txt
streamlit
ultralytics
transformers
torch
easyocr
opencv-python-headless
pillow
numpy
requests
beautifulsoup4
